Smell Code identified: Colors and Sizes enum moved in their own classes. The code impacted is the test files.
