+++ src/main/java/ch/heigvd/pdl/refactoring/Colors.java
@@ -1,0 +1,12 @@
+package ch.heigvd.pdl.refactoring;
+
+public enum Colors {
+    BLUE,
+    RED,
+    YELLOW,
+    NO_COLOR;
+    @Override
+    public String toString() {
+        return this.name().toLowerCase();
+    }
+}
