+++ src/main/java/ch/heigvd/pdl/refactoring/Sizes.java
@@ -1,0 +1,11 @@
+package ch.heigvd.pdl.refactoring;
+
+public enum Sizes {
+    XS,
+    S,
+    M,
+    L,
+    XL,
+    XXL,
+    SIZE_NOT_APPLICABLE
+}
