+++ src/test/java/ch/heigvd/pdl/refactoring/OrdersWriterTest.java
@@ -34,28 +34,28 @@ public class OrdersWriterTest {
 
     @Test
     public void OneOrderWithOneProduct() {
-        order111.addProduct(new Product("Shirt", Product.Colors.BLUE, Product.Sizes.M, 2.99, "TWD"));
+        order111.addProduct(new Product("Shirt", Colors.BLUE, Sizes.M, 2.99, "TWD"));
         String order111Json = JsonOrder111WithProduct("{\"code\":\"Shirt\",\"color\":\"blue\",\"size\":\"M\",\"price\":2.99,\"currency\":\"TWD\"}");
         assertEquals("{\"orders\":[" + order111Json + "]}", new OrdersWriter(orders).getContents());
     }
 
     @Test
     public void OneOrderWithOneProductNoSize() {
-        order111.addProduct(new Product("Pot", Product.Colors.RED, Product.Sizes.SIZE_NOT_APPLICABLE, 16.50, "SGD"));
+        order111.addProduct(new Product("Pot", Colors.RED, Sizes.SIZE_NOT_APPLICABLE, 16.50, "SGD"));
         String order111Json = JsonOrder111WithProduct("{\"code\":\"Pot\",\"color\":\"red\",\"price\":16.5,\"currency\":\"SGD\"}");
         assertEquals("{\"orders\":[" + order111Json + "]}", new OrdersWriter(orders).getContents());
     }
 
     @Test
     void CheckColors() {
-        Product product = new Product("Pot", Product.Colors.RED, Product.Sizes.SIZE_NOT_APPLICABLE, 16.50, "SGD");
+        Product product = new Product("Pot", Colors.RED, Sizes.SIZE_NOT_APPLICABLE, 16.50, "SGD");
         String color = "red";
         assertEquals(color, product.getColor());
     }
 
     @Test
     void CheckSizes() {
-        Product product = new Product("Pot", Product.Colors.RED, Product.Sizes.M, 16.50, "SGD");
+        Product product = new Product("Pot", Colors.RED, Sizes.M, 16.50, "SGD");
         String color = "M";
         assertEquals(color, product.getSizeName());
     }
